#51 - RefreshToken - Auto logout when token expired p4
