+++ client/src/app/(public)/(auth)/refresh-token/page.tsx
@@ -1,0 +1,22 @@
+'use client'
+
+import { checkAndRefreshToken, getRefreshTokenFromLocalStorage } from '@/lib/utils'
+import { useRouter, useSearchParams } from 'next/navigation'
+import React, { useEffect } from 'react'
+
+export default function RefreshTokenPage() {
+  const router = useRouter()
+  const searchParams = useSearchParams()
+  const refreshTokenFromUrl = searchParams.get('refreshToken')
+  const redirectPathName = searchParams.get('redirect')
+  useEffect(() => {
+    if (refreshTokenFromUrl && refreshTokenFromUrl === getRefreshTokenFromLocalStorage()) {
+      checkAndRefreshToken({
+        onSuccess: () => {
+          router.push(redirectPathName || '')
+        }
+      })
+    }
+  }, [router, refreshTokenFromUrl, redirectPathName])
+  return <div>Refresh token...</div>
+}
