+++ client/src/middleware.ts
@@ -22,8 +22,9 @@ export function middleware(request: NextRequest) {
 
   // Trường hơp đăng nhập rồi nhưng access token lại hết hạn
   if (privatePaths.some((path) => pathname.startsWith(path) && !accessToken && refreshToken)) {
-    const url = new URL('/logout', request.url)
+    const url = new URL('/refresh-token', request.url)
     url.searchParams.set('refreshToken', refreshToken ?? '')
+    url.searchParams.set('redirect', pathname)
     return NextResponse.redirect(url)
   }
